# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for XuXaDex Platform.",
        "title": "XuXaDex API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/match/all": {
            "get": {
                "description": "Get Quick Matches handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match"
                ],
                "summary": "Get Quick Matches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "id:ASC",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "count\u003c100",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_match.MatchEntity"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/match/quick": {
            "post": {
                "description": "Create Quick Match handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "match"
                ],
                "summary": "Create Quick Match",
                "parameters": [
                    {
                        "description": "quick match create data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_match.QuickMatchCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_match.MatchEntity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_pkg_responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_xuxadex_backend-mvp-main_internal_match.MatchEntity": {
            "type": "object",
            "properties": {
                "creator": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "team1": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_team.TeamEntity"
                },
                "team1_score": {
                    "type": "integer"
                },
                "team2": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_team.TeamEntity"
                },
                "team2_score": {
                    "type": "integer"
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_internal_match.QuickMatchCreateDTO": {
            "type": "object",
            "properties": {
                "creator": {
                    "type": "string"
                },
                "max_players": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_internal_player.PlayerEntity": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_user.UserBaseEntity"
                }
            }
        },
        "github_com_xuxadex_backend-mvp-main_internal_team.TeamEntity": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "max_players": {
                    "type": "integer"
                },
                "min_players": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xuxadex_backend-mvp-main_internal_player.PlayerEntity"
                    }
                }
            }
        },
...
        "github_com_xuxadex_backend-mvp-main_internal_user.UserBaseEntity": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "github_com_xuxadex_backend-mvp-main_pkg_responses.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    }
}